requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.6.0
beanie==1.30.0
certifi==2025.7.14
cffi==1.17.1
click==8.2.1
cloudinary==1.44.1
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
fastapi==0.116.1
h11==0.16.0
idna==3.10
lazy-model==0.2.0
motor==3.7.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pymongo==4.13.2
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
Pillow==11.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
